Dev (#1)
* styles: primevue, tailwindcss

* rm defaults

* feat: Add common type declarations

* chore: Update npm dependencies for Vue and Vue Router

* update router

* min nav_machine

* refactor

* update nav_machine

* define auth_machine

* problem with initialization order of nav_machine and router...

* fix bootstrap initialization of references error

* fix: declare NavEv_global type

* feat: Add RootLayout component for improved page structure

* xstate-vue-router integration

* refactor, start implement on page user role with optional dynamic nik

* viewer role on page correction

* local_storage utils, openapi

* add api

* add vue3-google-login

* add fetcher_machine

* add fetcher_machine

* fetch, auth, nav, nik => it works!

* fix initial auth

* update RouterLink :to

* RootLayout update

* feat: Add "Settings" link to navigation menu

The code changes include adding a new list item to the navigation menu in the `RootLayout.vue` file, which links to the "Settings" page. Additionally, the `types.x.nav.d.ts` file was modified to include a new navigation type for the "Settings" page. The `navigate.action.ts` file was also added to handle the navigation action for the "Settings" page. Finally, the `router.ts` file was updated to include a new route for the "Settings" page.

Based on the recent user commits and repository commits, it seems that the developer has been working on various aspects of the application, including updates to the layout, authentication, fetching data, and adding new features like the Google login and API integration.

* autologout

* fix: restrice first hardcoded navigate to private url by guest

* PageSettings

* chore: Update font-family in main.css and add PrimeVue components

The main.css file was updated to change the font-family to 'Courier New', Courier, monospace. Additionally, the PrimeVue components InputText and FloatLabel were added in the main.ts file.

* ts fix

* page_settings.machine (begin)

* fix: Update viewer role to 'owner' when necessary

The code changes in raise_nav_ev.action.ts update the viewer role to 'owner' in certain conditions. This ensures that the viewer role is set correctly based on the provided values.

* explore how to integrate machine with ui

* split logic into many pieces

* it works... but missing conventions, code will be messi soon!

* pause

---------

Co-authored-by: nik-kita <nik-kita.deno.dev>